requirements.txt
snowflake-connector-python==3.9.1
snowflake-snowpark-python==1.14.0